requirements.txt
fastapi
uvicorn
google-cloud-firestore
firebase-admin